+++ mayan/apps/templating/static/templating/js/template_widget.js
@@ -25,4 +25,28 @@ jQuery(document).ready(function() {
         );
         $(fieldNameID + ' option')[0].selected = true;
     });
+
+    $('[data-autocopy="true"]').change(function(event) {
+        var $this = $(this);
+        var $idTemplate = $('[data-template-fields="template"]');
+        var templateCursorPosition = $idTemplate.prop('selectionStart');
+        var templateValue = $idTemplate.val();
+        var fieldText = eval('`' + $this.data('field-template') +'`');
+
+        templateValue = templateValue.slice(
+            0, templateCursorPosition
+        ) + fieldText + templateValue.slice(
+            templateCursorPosition
+        );
+        $idTemplate.val(templateValue);
+        $idTemplate.focus();
+        $idTemplate.prop(
+            'selectionStart', templateCursorPosition + fieldText.length
+        );
+        $idTemplate.prop(
+            'selectionEnd', templateCursorPosition + fieldText.length
+        );
+
+        $this.val('');
+    });
 });